refactor(progress/rating): optimize CSS
