+++ src/components/progress/themes/circular/shared/circular.progress.bootstrap.scss
@@ -1,15 +1,8 @@
 @use 'styles/utilities' as *;
-@use '../light/themes' as *;
-
-$theme: $bootstrap;
 
 :host {
     --stroke-thickness: 4px;
     --diameter: 28px;
-}
-
-[part~='track'] {
-    stroke: var-get($theme, 'base-circle-color');
 }
 
 [part~='svg'][part~='indeterminate'] {
@@ -24,33 +17,3 @@ $theme: $bootstrap;
 [part~='track'][part~='indeterminate'] {
     stroke: transparent;
 }
-
-[part~='label'] {
-    fill: var-get($theme, 'text-color');
-}
-
-[part~='gradient_start'],
-[part~='gradient_end'] {
-    stop-color: var-get($theme, 'progress-circle-color');
-}
-
-:host([variant='danger']) [part~='gradient_start'],
-:host([variant='danger']) [part~='gradient_end'] {
-    stop-color: color(error, 500);
-}
-
-:host([variant='warning']) [part~='gradient_start'],
-:host([variant='warning']) [part~='gradient_end'] {
-    stop-color: color(warn, 500);
-}
-
-:host([variant='info']) [part~='gradient_start'],
-:host([variant='info']) [part~='gradient_end'] {
-    stop-color: color(info, 500);
-}
-
-:host([variant='success']) [part~='gradient_start'],
-:host([variant='success']) [part~='gradient_end'] {
-    stop-color: color(success, 500);
-}
-
