+++ src/components/progress/themes/circular/shared/circular.progress.common.scss
