+++ src/components/progress/themes/circular/shared/circular.progress.fluent.scss
@@ -1,43 +1,7 @@
 @use 'styles/utilities' as *;
-@use '../light/themes' as *;
 
 // TODO If possible make the animation the same as the MS fluent
 
-$theme: $fluent;
-
 :host {
     --stroke-thickness: 1.5px;
-}
-
-[part~='track'] {
-    stroke: var-get($theme, 'base-circle-color');
-}
-
-[part~='label'] {
-    fill: var-get($theme, 'text-color');
-}
-
-[part~='gradient_start'],
-[part~='gradient_end'] {
-    stop-color: var-get($theme, 'progress-circle-color');
-}
-
-:host([variant='danger']) [part~='gradient_start'],
-:host([variant='danger']) [part~='gradient_end'] {
-    stop-color: color(error, 500);
-}
-
-:host([variant='warning']) [part~='gradient_start'],
-:host([variant='warning']) [part~='gradient_end'] {
-    stop-color: color(warn, 500);
-}
-
-:host([variant='info']) [part~='gradient_start'],
-:host([variant='info']) [part~='gradient_end'] {
-    stop-color: color(info, 500);
-}
-
-:host([variant='success']) [part~='gradient_start'],
-:host([variant='success']) [part~='gradient_end'] {
-    stop-color: color(success, 500);
 }
