+++ src/components/progress/themes/circular/themes.ts
@@ -12,38 +12,37 @@ import { styles as indigoLight } from './light/circular.progress.indigo.css.js';
 import { styles as materialLight } from './light/circular.progress.material.css.js';
 // Shared Styles
 import { styles as bootstrap } from './shared/circular.progress.bootstrap.css.js';
+import { styles as shared } from './shared/circular.progress.common.css.js';
 import { styles as fluent } from './shared/circular.progress.fluent.css.js';
-import { styles as indigo } from './shared/circular.progress.indigo.css.js';
-import { styles as material } from './shared/circular.progress.material.css.js';
 import { Themes } from '../../../../theming/types.js';
 
 const light = {
   bootstrap: css`
-    ${bootstrap} ${bootstrapLight}
+    ${shared} ${bootstrap} ${bootstrapLight}
   `,
   material: css`
-    ${material} ${materialLight}
+    ${shared} ${materialLight}
   `,
   fluent: css`
-    ${fluent} ${fluentLight}
+    ${shared} ${fluent} ${fluentLight}
   `,
   indigo: css`
-    ${indigo} ${indigoLight}
+    ${shared} ${indigoLight}
   `,
 };
 
 const dark = {
   bootstrap: css`
-    ${bootstrap} ${bootstrapDark}
+    ${shared} ${bootstrap} ${bootstrapDark}
   `,
   material: css`
-    ${material} ${materialDark}
+    ${shared} ${materialDark}
   `,
   fluent: css`
-    ${fluent} ${fluentDark}
+    ${shared} ${fluent} ${fluentDark}
   `,
   indigo: css`
-    ${indigo} ${indigoDark}
+    ${shared} ${indigoDark}
   `,
 };
