+++ src/components/progress/themes/linear/shared/linear.progress.common.scss
