+++ src/components/progress/themes/linear/shared/linear.progress.fluent.scss
@@ -18,28 +18,12 @@ $track-color: color(gray, 200);
 }
 
 [part~='track'] {
-    border-radius: var-get($theme, 'track-border-radius');
-    background: var-get($theme, 'track-color');
     margin: rem(8px) 0;
 }
 
 [part~='label'] {
     font-size: rem(14px);
     font-weight: 400;
-}
-
-[part~='value'] {
-    color: var-get($theme, 'text-color')
-}
-
-[part~='primary']:not([part~='indeterminate']) {
-    background-color: var-get($theme, 'fill-color-default');
-}
-
-[part~='primary'] {
-    &::after {
-        background-color: var-get($theme, 'fill-color-default');
-    }
 }
 
 [part~='indeterminate']:not([part~='secondary']) {
@@ -78,94 +62,4 @@ $track-color: color(gray, 200);
 
 [part~='success'][part~='indeterminate'] {
     background: linear-gradient(90deg, transparent 0%, var-get($theme, 'fill-color-success') 50%, transparent 100%);
-}
-
-[part~='striped'][part~='primary']:not([part~='indeterminate']) {
-    @include stripe-colors(var-get($theme, 'fill-color-default'), var-get($theme, 'stripes-color'), $stripe-size, -45deg);
-}
-
-[part~='striped'][part~='primary'] {
-    &::after {
-        @include stripe-colors(var-get($theme, 'fill-color-default'), var-get($theme, 'stripes-color'), $stripe-size, -45deg);
-    }
-}
-
-[part~='danger']:not([part~='indeterminate']) {
-    background-color: var-get($theme, 'fill-color-danger');
-}
-
-[part~='danger'] {
-    &::after {
-        background-color: var-get($theme, 'fill-color-danger');
-    }
-}
-
-[part~='striped'][part~='danger']:not([part~='indeterminate']) {
-    @include stripe-colors(var-get($theme, 'fill-color-danger'), var-get($theme, 'stripes-color'), $stripe-size, -45deg);
-}
-
-[part~='striped'][part~='danger'] {
-    &::after {
-        @include stripe-colors(var-get($theme, 'fill-color-danger'), var-get($theme, 'stripes-color'), $stripe-size, -45deg);
-    }
-}
-
-[part~='warning']:not([part~='indeterminate']) {
-    background-color: var-get($theme, 'fill-color-warning');
-}
-
-[part~='warning'] {
-    &::after {
-        background-color: var-get($theme, 'fill-color-warning');
-    }
-}
-
-[part~='striped'][part~='warning']:not([part~='indeterminate']) {
-    @include stripe-colors(var-get($theme, 'fill-color-warning'), var-get($theme, 'stripes-color'), $stripe-size, -45deg);
-}
-
-[part~='striped'][part~='warning'] {
-    &::after {
-        @include stripe-colors(var-get($theme, 'fill-color-warning'), var-get($theme, 'stripes-color'), $stripe-size, -45deg);
-    }
-}
-
-[part~='info']:not([part~='indeterminate']) {
-    background-color: var-get($theme, 'fill-color-info');
-}
-
-[part~='info'] {
-    &::after {
-        background-color: var-get($theme, 'fill-color-info');
-    }
-}
-
-[part~='striped'][part~='info']:not([part~='indeterminate']) {
-    @include stripe-colors(var-get($theme, 'fill-color-info'), var-get($theme, 'stripes-color'), $stripe-size, -45deg);
-}
-
-[part~='striped'][part~='info'] {
-    &::after {
-        @include stripe-colors(var-get($theme, 'fill-color-info'), var-get($theme, 'stripes-color'), $stripe-size, -45deg);
-    }
-}
-
-[part~='success']:not([part~='indeterminate']) {
-    background-color: var-get($theme, 'fill-color-success');
-}
-
-[part~='success'] {
-    &::after {
-        background-color: var-get($theme, 'fill-color-success');
-    }
-}
-
-[part~='striped'][part~='success']:not([part~='indeterminate']) {
-    @include stripe-colors(var-get($theme, 'fill-color-success'), var-get($theme, 'stripes-color'), $stripe-size, -45deg);
-}
-
-[part~='striped'][part~='success'] {
-    &::after {
-        @include stripe-colors(var-get($theme, 'fill-color-success'), var-get($theme, 'stripes-color'), $stripe-size, -45deg);
-    }
 }
