+++ src/components/progress/themes/linear/shared/linear.progress.indigo.scss
@@ -7,11 +7,6 @@ $theme: $indigo;
 $stripe-size: rem(6px);
 $stripe-space: rem(14px);
 
-[part~='track'] {
-    border-radius: var-get($theme, 'track-border-radius');
-    background: var-get($theme, 'track-color');
-}
-
 [part~='striped']:not([part~='indeterminate']) {
     /* stylelint-disable */
     background: var-get($theme, 'fill-color-default') repeating-linear-gradient(
@@ -22,109 +17,4 @@ $stripe-space: rem(14px);
         var-get($theme, 'stripes-color') $stripe-space
     );
     /* stylelint-enable */
-}
-
-
-[part~='value'] {
-    color: var-get($theme, 'text-color')
-}
-
-[part~='primary']:not([part~='indeterminate']) {
-    background-color: var-get($theme, 'fill-color-default');
-}
-
-[part~='primary'] {
-    &::after {
-        background-color: var-get($theme, 'fill-color-default');
-    }
-}
-
-[part~='striped'][part~='primary']:not([part~='indeterminate']) {
-    @include stripe-colors(var-get($theme, 'fill-color-default'), var-get($theme, 'stripes-color'), $stripe-size, -45deg);
-}
-
-[part~='striped'][part~='primary'] {
-    &::after {
-        @include stripe-colors(var-get($theme, 'fill-color-default'), var-get($theme, 'stripes-color'), $stripe-size, -45deg);
-    }
-}
-
-[part~='danger']:not([part~='indeterminate']) {
-    background-color: var-get($theme, 'fill-color-danger');
-}
-
-[part~='danger'] {
-    &::after {
-        background-color: var-get($theme, 'fill-color-danger');
-    }
-}
-
-[part~='striped'][part~='danger']:not([part~='indeterminate']) {
-    @include stripe-colors(var-get($theme, 'fill-color-danger'), var-get($theme, 'stripes-color'), $stripe-size, -45deg);
-}
-
-[part~='striped'][part~='danger'] {
-    &::after {
-        @include stripe-colors(var-get($theme, 'fill-color-danger'), var-get($theme, 'stripes-color'), $stripe-size, -45deg);
-    }
-}
-
-[part~='warning']:not([part~='indeterminate']) {
-    background-color: var-get($theme, 'fill-color-warning');
-}
-
-[part~='warning'] {
-    &::after {
-        background-color: var-get($theme, 'fill-color-warning');
-    }
-}
-
-[part~='striped'][part~='warning']:not([part~='indeterminate']) {
-    @include stripe-colors(var-get($theme, 'fill-color-warning'), var-get($theme, 'stripes-color'), $stripe-size, -45deg);
-}
-
-[part~='striped'][part~='warning'] {
-    &::after {
-        @include stripe-colors(var-get($theme, 'fill-color-warning'), var-get($theme, 'stripes-color'), $stripe-size, -45deg);
-    }
-}
-
-[part~='info']:not([part~='indeterminate']) {
-    background-color: var-get($theme, 'fill-color-info');
-}
-
-[part~='info'] {
-    &::after {
-        background-color: var-get($theme, 'fill-color-info');
-    }
-}
-
-[part~='striped'][part~='info']:not([part~='indeterminate']) {
-    @include stripe-colors(var-get($theme, 'fill-color-info'), var-get($theme, 'stripes-color'), $stripe-size, -45deg);
-}
-
-[part~='striped'][part~='info'] {
-    &::after {
-        @include stripe-colors(var-get($theme, 'fill-color-info'), var-get($theme, 'stripes-color'), $stripe-size, -45deg);
-    }
-}
-
-[part~='success']:not([part~='indeterminate']) {
-    background-color: var-get($theme, 'fill-color-success');
-}
-
-[part~='success'] {
-    &::after {
-        background-color: var-get($theme, 'fill-color-success');
-    }
-}
-
-[part~='striped'][part~='success']:not([part~='indeterminate']) {
-    @include stripe-colors(var-get($theme, 'fill-color-success'), var-get($theme, 'stripes-color'), $stripe-size, -45deg);
-}
-
-[part~='striped'][part~='success'] {
-    &::after {
-        @include stripe-colors(var-get($theme, 'fill-color-success'), var-get($theme, 'stripes-color'), $stripe-size, -45deg);
-    }
 }
