+++ src/components/rating/themes/shared/rating.common.scss
@@ -1,4 +1,4 @@
-@use '../../../../styles/utilities/index' as *;
+@use 'styles/utilities/index' as *;
 @use '../light/themes' as *;
 
 $theme: $material;
