+++ src/components/rating/themes/shared/rating.fluent.scss
@@ -1,12 +1,4 @@
 @use 'styles/utilities/index' as *;
-@use '../light/themes' as *;
-
-$theme: $fluent;
-
-:host {
-    --symbol-full-color: #{var-get($theme, 'symbol-full-color')};
-    --symbol-empty-color: #{var-get($theme, 'symbol-empty-color')};
-}
 
 :host,
 [part='base'] {
@@ -15,29 +7,8 @@ $theme: $fluent;
 
 [part~='label'] {
     @include type-style('subtitle-2');
-
-    color: var-get($theme, 'label-color');
 }
 
 [part~='value-label'] {
     @include type-style('body-2');
-
-    color: var-get($theme, 'value-label');
 }
-
-:host([disabled]),
-:host(:disabled) {
-    --symbol-full-color: #{var-get($theme, 'disabled-full-symbol-color')};
-    --symbol-empty-color: #{var-get($theme, 'disabled-empty-symbol-color')};
-    --disabled-symbol-full-filter: #{var-get($theme, 'symbol-full-filter')};
-    --disabled-symbol-empty-filter: #{var-get($theme, 'symbol-empty-filter')};
-
-    [part~='label'] {
-        color: var-get($theme, 'disabled-label-color');
-    }
-
-    [part~='value-label'] {
-        color: var-get($theme, 'disabled-label-color');
-    }
-}
-
