+++ src/components/rating/themes/shared/rating.indigo.scss
@@ -1,12 +1,6 @@
 @use 'styles/utilities/index' as *;
-@use '../light/themes' as *;
-
-$theme: $indigo;
 
 :host {
-    --symbol-full-color: #{var-get($theme, 'symbol-full-color')};
-    --symbol-empty-color: #{var-get($theme, 'symbol-empty-color')};
-
     gap: rem(5px);
 }
 
@@ -17,28 +11,4 @@ $theme: $indigo;
 [part~='label'],
 [part~='value-label'] {
     @include type-style('body-1');
-}
-
-[part~='label'] {
-    color: var-get($theme, 'label-color');
-}
-
-[part~='value-label'] {
-    color: var-get($theme, 'value-label');
-}
-
-:host([disabled]),
-:host(:disabled) {
-    --symbol-full-color: #{var-get($theme, 'disabled-full-symbol-color')};
-    --symbol-empty-color: #{var-get($theme, 'disabled-empty-symbol-color')};
-    --disabled-symbol-full-filter: #{var-get($theme, 'symbol-full-filter')};
-    --disabled-symbol-empty-filter: #{var-get($theme, 'symbol-empty-filter')};
-
-    [part~='label'] {
-        color: var-get($theme, 'disabled-label-color');
-    }
-
-    [part~='value-label'] {
-        color: var-get($theme, 'disabled-label-color');
-    }
 }
